requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
cloudpickle==1.6.0; python_version >= '3.5'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gym-retro==0.8.0
gym==0.18.0; python_version >= '3.6'
neat-python==0.92
numpy==1.20.2; python_version >= '3.7'
opencv-python==4.5.1.48
pillow==7.2.0; python_version >= '3.5'
pyglet==1.5.0
scipy==1.6.2; python_version < '3.10' and python_version >= '3.7'
